[level 0] Title: 옷가게 할인 받기, Time: 0.03 ms, Memory: 77.9 MB -BaekjoonHub

+++ 프로그래머스/0/120818. 옷가게 할인 받기/옷가게 할인 받기.java
@@ -1,5 +1,5 @@
 class Solution {
     public int solution(int price) {
-        return (price >= 500000) ? (int)((double)price * 0.8) : (price >= 300000) ? (int)((double)price * 0.9) : (price >= 100000) ? (int)((double)price * 0.95) : price;
+        return (price >= 500000) ? (int)(price * 0.8) : (price >= 300000) ? (int)(price * 0.9) : (price >= 100000) ? (int)(price * 0.95) : price;
     }
 }
